refactor: Move rental detail UI models to presentation layer (#203)
- Moved `PriceSummaryUiModel`, `RentalSummaryUiModel`, and `RequestAcceptDialogUiModel` from `common.uimodel` to `presentation.rentaldetail.model`
- Updated import paths accordingly
- Modified `RequestAcceptDialog` to accept individual parameters to prevent circular dependency caused by the model relocation
