+++ app/src/main/java/com/example/rentit/presentation/rentaldetail/model/PriceSummaryUiModel.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.common.uimodel
+package com.example.rentit.presentation.rentaldetail.model
 
 data class PriceSummaryUiModel(
     val label: String,
